Solved bug autoconnect
