+++ dapp/pages/_app.tsx
@@ -2,10 +2,16 @@ import '../styles/globals.css'
 import type { AppProps } from 'next/app'
 import { Layout } from '../components'
 import { ThemeProvider } from 'next-themes'
-import { DAppProvider } from '@usedapp/core'
+import { DAppProvider, Localhost } from '@usedapp/core'
+
 import { Provider as CeramicProvider } from '@self.id/react'
 
-const config = {}
+const config = {
+  readOnlyChainId: Localhost.chainId,
+  readOnlyUrls: {
+    [Localhost.chainId]: 'http://127.0.0.1:8545',
+  },
+}
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
